capitalizeFirst splitted
