+++ interface/src/js/capitalizeFirst.js
@@ -1,0 +1,2 @@
+export const capitalizeFirst = (str, lower = false) =>
+    (lower ? str.toLowerCase() : str).replace(/(?:^|\s|-|["'([{])+\S/g, match => match.toUpperCase());
